backend/omi-discord-bot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.2
cffi==2.0.0
discord-py==2.6.3
frozenlist==1.7.0
idna==3.10
joblib==1.5.2
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.6.4
numpy==2.3.3
propcache==0.3.2
pycparser==2.23
pygments==2.19.2
pynacl==1.6.0
python-dotenv==1.1.1
rank-bm25==0.2.2
rapidfuzz==3.14.1
rich==14.1.0
scikit-learn==1.7.2
scipy==1.16.2
thefuzz==0.22.1
threadpoolctl==3.6.0
watchdog==6.0.0
yarl==1.20.1
